requirements.txt
torch
torchvision
addict
opencv-python
Pillow
numpy